parcel_tracker_lib/requirements.txt
gevent==1.2.0
greenlet==0.4.11
lxml==3.7.1
python-dateutil==2.6.0
six==1.10.0
